fix: Handle the fact that the QHull library has an upper limit on the number of points it can handle
In cases where the number of points exceeds this limit, optionally allow subsampling of the given points.

+++ source/qhull.cpp
@@ -25,13 +25,13 @@
 
 extern "C" 
 {
-  orgQhull::Qhull * convexHullConstructorC(long n, double *points[], int *status);
+  orgQhull::Qhull * convexHullConstructorC(int n, double *points[], int *status);
   void convexHullDestructorC(orgQhull::Qhull *qhull);
   double convexHullVolumeC(orgQhull::Qhull *qhull);
   bool convexHullPointIsInHullC(orgQhull::Qhull *qhull, double point[]);
 }
 
-orgQhull::Qhull * convexHullConstructorC(long n, double *points[], int *status)
+orgQhull::Qhull * convexHullConstructorC(int n, double *points[], int *status)
 {
   //% Constructor for convex hull objects.
